Allow a element for the target
