+++ src/vs/workbench/contrib/terminal/browser/widgets/environmentVariableInfoWidget.ts
@@ -7,7 +7,6 @@ import { Widget } from 'vs/base/browser/ui/widget';
 import { IEnvironmentVariableInfo } from 'vs/workbench/contrib/terminal/common/environmentVariable';
 import { MarkdownString } from 'vs/base/common/htmlContent';
 import { ITerminalWidget } from 'vs/workbench/contrib/terminal/browser/widgets/widgets';
-import { IHoverTarget, IHoverAnchor } from 'vs/workbench/contrib/hover/browser/hover';
 import { IInstantiationService } from 'vs/platform/instantiation/common/instantiation';
 import { HoverWidget } from 'vs/workbench/contrib/hover/browser/hoverWidget';
 import { RunOnceScheduler } from 'vs/base/common/async';
@@ -83,37 +82,10 @@ export class EnvironmentVariableInfoWidget extends Widget implements ITerminalWi
 		if (!this._domNode || !this._container || this._hoverWidget) {
 			return;
 		}
-		const target = new ElementHoverTarget(this._domNode);
 		const actions = this._info.getActions ? this._info.getActions() : undefined;
-		this._hoverWidget = this._instantiationService.createInstance(HoverWidget, target, new MarkdownString(this._info.getInfo()), () => { }, actions);
+		this._hoverWidget = this._instantiationService.createInstance(HoverWidget, this._domNode, new MarkdownString(this._info.getInfo()), () => { }, actions);
 		this._hoverWidget.render(this._container);
 		this._register(this._hoverWidget);
 		this._register(this._hoverWidget.onDispose(() => this._hoverWidget = undefined));
-	}
-}
-
-class ElementHoverTarget implements IHoverTarget {
-	readonly targetElements: readonly HTMLElement[];
-
-	constructor(
-		private _element: HTMLElement
-	) {
-		this.targetElements = [this._element];
-	}
-
-	get anchor(): IHoverAnchor {
-		const position = dom.getDomNodePagePosition(this._element);
-		return {
-			x: position.left,
-			y: document.documentElement.clientHeight - position.top - 1,
-			fallbackY: position.top + position.height
-		};
-	}
-
-	render(container: HTMLElement): void {
-		// TODO: Should render be optional?
-	}
-
-	dispose(): void {
 	}
 }
